+++ src/app/store.tsx
@@ -1,30 +1,16 @@
 import { configureStore } from '@reduxjs/toolkit';
 import { TypedUseSelectorHook, useDispatch, useSelector } from 'react-redux';
 import { listenerMiddleware } from './listenerMiddleware';
-import sampleSummaryReducer from '../components/Widgets/SampleSummary/sampleSummarySlice';
-import organisationsReducer from '../components/Widgets/Organisations/organisationsSlice';
-import projectDashboardReducer from '../components/Dashboards/ProjectDashboard/projectDashboardSlice';
-import stCountsReducer from '../components/Widgets/StCounts/stCountsSlice';
-import thresholdAlertsReducer from '../components/Widgets/ThresholdAlerts/thresholdAlertsSlice';
-import qcStatusReducer from '../components/Widgets/QcStatus/qcStatusSlice';
-import phessIdStatusReducer from '../components/Widgets/PhessIdStatus/phessIdStatusSlice';
 import userDashboardReducer from '../components/Dashboards/UserDashboard/userDashboardSlice';
-import userOverviewReducer from '../components/Widgets/UserOverview/userOverviewSlice';
-import projectsTotalReducer from '../components/Widgets/ProjectsTotal/projectsTotalSlice';
-import phessIdOverallReducer from '../components/Widgets/PhessIdOverall/phessIdOverallSlice';
+import userOverviewReducer from '../components/Widgets/UserWidgets/UserOverview/userOverviewSlice';
+import projectsTotalReducer from '../components/Widgets/UserWidgets/ProjectsTotal/projectsTotalSlice';
+import phessIdOverallReducer from '../components/Widgets/UserWidgets/PhessIdOverall/phessIdOverallSlice';
 import projectMetadataReducer from './projectMetadataSlice';
 import groupMetadataReducer from './groupMetadataSlice';
 import userReducer from './userSlice';
 
 const store = configureStore({
   reducer: {
-    sampleSummaryState: sampleSummaryReducer,
-    organisationsState: organisationsReducer,
-    qcStatusState: qcStatusReducer,
-    phessIdStatusState: phessIdStatusReducer,
-    stCountsState: stCountsReducer,
-    thresholdAlertsState: thresholdAlertsReducer,
-    projectDashboardState: projectDashboardReducer,
     userDashboardState: userDashboardReducer,
     userOverviewState: userOverviewReducer,
     projectTotalState: projectsTotalReducer,